compiler.go: assembly parser label fixes.
EVM uses big-endian byte-order, so to pad a label value to 4 bytes,
zeros must be added to the front, not the end.

Incrementing PC by 5 is only correct if the label appears after a jump,
in which case there is an implicit push. When it appears after an explicit
push, PC should only be incremented by 4.

This way, a label can be pushed explicitly, or loaded from memory to
implement a jump table.